requirements.txt
pyspark>=3.4,<4.0
pytest>=7.0
python-dotenv>=1.0
